agents/hotel_agent/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents
mcp
requests>=2.31.0
httpx>=0.24.0
